# specs/001-note-management-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Quick Note — Note Management API
  version: 1.0.0
  summary: Personal note capture, organization, search, and recovery.
  description: |
    This document is the authoritative contract for the Quick Note service. It is hand-authored and
    precedes the implementation, per Constitution Principle I. Where the running service and this
    document disagree, this document is correct and the service is defective.

    **Authentication.** Every operation requires a Bearer JWT issued by the configured Keycloak
    realm. The service verifies signature, `iss`, `aud`, and `exp`, and derives the acting user from
    the `sub` claim. It never issues, refreshes, or revokes tokens. A user identifier supplied
    anywhere else in a request is ignored.

    **Errors.** Every non-2xx response is `application/problem+json` per RFC 9457, carrying a stable
    machine-readable `code`. Codes are part of this contract; see `contracts/error-catalog.md`.

    **Non-disclosure.** A resource that does not exist and a resource owned by another user both
    return `404`. The API never reveals which of the two applies.
servers:
  - url: https://api.quicknote.example/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /trash:
    get:
      tags: [Trash]
      operationId: listTrash
      summary: List trashed notes
      description: Ordered by `trashedAt`, most recently trashed first.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: A page of trashed notes.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotePage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    delete:
      tags: [Trash]
      operationId: emptyTrash
      summary: Empty the trash
      description: |
        Permanently and irreversibly removes every trashed note. Active and archived notes are
        unaffected (FR-023). Labels are not deleted.
      responses:
        '204': { description: Trash is empty. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /trash/{noteId}:
    parameters:
      - $ref: '#/components/parameters/NoteId'
    delete:
      tags: [Trash]
      operationId: purgeNote
      summary: Permanently delete a trashed note
      description: |
        Irreversible (FR-025). To recover a note instead, use
        `PATCH /notes/{noteId}` with `{"trashed": false}`.
      responses:
        '204': { description: Note permanently deleted. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: The note does not exist, is not owned by the caller, or is not in the trash.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by the configured Keycloak realm. The service verifies signature, `iss`, `aud`,
        and `exp`, and reads the user identifier from `sub`.
  parameters:
    NoteId:
      name: noteId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      required: false
      description: Page size. Requests above the maximum are rejected, not clamped (FR-039).
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
  schemas:
    NoteColor:
      type: string
      enum: [default, red, orange, yellow, green, teal, blue, dark_blue, purple, pink, brown]
      default: default
    Note:
      type: object
      additionalProperties: false
      required: [id, color, pinned, archived, trashed, labelIds, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        title:
          type: [string, 'null']
          maxLength: 200
        body:
          type: [string, 'null']
          maxLength: 20000
        color: { $ref: '#/components/schemas/NoteColor' }
        pinned:
          type: boolean
          description: Never true while `archived` or `trashed` is true (FR-020).
        archived:
          type: boolean
          description: |
            A note is in exactly one of three states — active, archived, or trashed. `archived` and
            `trashed` are the wire form of that single state and are never both true; both false
            means the note is active.
        trashed:
          type: boolean
          description: See `archived`. Never true at the same time as `archived`.
        trashedAt:
          type: [string, 'null']
          format: date-time
          description: Non-null if and only if `trashed` is true.
        labelIds:
          type: array
          maxItems: 20
          uniqueItems: true
          items: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    NotePage:
      type: object
      additionalProperties: false
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Note' }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        totalElements: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        nextPage:
          type: [string, 'null']
          format: uri-reference
          description: URI of the next page, or null on the last page.
...
    ProblemDetail:
      type: object
      description: RFC 9457 Problem Details, extended with a stable machine-readable `code`.
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail:
          type: string
          description: Human-readable. Never contains stack traces, SQL, or upstream error text (FR-043).
        instance: { type: string, format: uri-reference }
        code:
          type: string
          description: Stable identifier; part of this contract. See `error-catalog.md`.
        correlationId:
          type: string
          description: Echoes the request correlation ID, for support and log lookup.
        errors:
          type: array
          description: Present on validation failures; one entry per offending field (FR-041).
          items:
            type: object
            required: [field, code, message]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
  responses:
    BadRequest:
      description: Malformed request or failed validation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Unauthorized:
      description: |
        Missing, expired, malformed, or unverifiable credential. The response does not reveal which
        check failed (FR-003).
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
    TooManyRequests:
      description: Per-principal or per-IP rate limit exceeded.
      headers:
        Retry-After:
          schema: { type: integer }
        RateLimit-Remaining:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
    InternalError:
      description: |
        An unhandled fault. The body carries only `type`, `title`, `status`, `code`, and
        `correlationId` — never a stack trace, SQL fragment, hostname, or upstream error text
        (FR-043). Full detail is logged against the same correlation ID.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    ServiceUnavailable:
      description: |
        A critical dependency is unavailable. This is also what the readiness probe reflects.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }